[Enterprise Search] Group Connector Advanced Config together (#168533)
## Summary

Advanced Connector configuration from Rich Configurable Fields are
grouped together.

<img width="707" alt="Screenshot 2023-10-10 at 22 21 52"
src="https://github.com/elastic/kibana/assets/1410658/469d4f24-26ac-442b-acae-73f69ab8a7eb">
<img width="747" alt="Screenshot 2023-10-10 at 22 22 03"
src="https://github.com/elastic/kibana/assets/1410658/c4b56bde-ebec-403c-84ef-55a770cb3b89">

### Checklist

Delete any items that are not applicable to this PR.

- [x] Any text added follows [EUI's writing
guidelines](https://elastic.github.io/eui/#/guidelines/writing), uses
sentence case text and includes [i18n
support](https://github.com/elastic/kibana/blob/main/packages/kbn-i18n/README.md)
- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [x] Any UI touched in this PR is usable by keyboard only (learn more
about [keyboard accessibility](https://webaim.org/techniques/keyboard/))
- [x] This renders correctly on smaller devices using a responsive
layout. (You can test this [in your
browser](https://www.browserstack.com/guide/responsive-testing-on-local-server))
- [x] This was checked for [cross-browser
compatibility](https://www.elastic.co/support/matrix#matrix_browsers)

---------

(cherry picked from commit 2df247a75b97a7c538bea3993876e9e98418ba7a)
