+++ x-pack/plugins/enterprise_search/public/applications/enterprise_search_content/components/search_index/connector/connector_configuration_logic.ts
@@ -7,14 +7,10 @@
 
 import { kea, MakeLogicType } from 'kea';
 
-import { i18n } from '@kbn/i18n';
-
 import {
   ConnectorConfigProperties,
   ConnectorConfiguration,
   ConnectorStatus,
-  Dependency,
-  FieldType,
 } from '@kbn/search-connectors';
 
 import { isCategoryEntry } from '../../../../../../common/connectors/is_category_entry';
@@ -30,6 +26,11 @@ import {
 } from '../../../api/index/cached_fetch_index_api_logic';
 import { FetchIndexApiResponse } from '../../../api/index/fetch_index_api_logic';
 import { isConnectorIndex } from '../../../utils/indices';
+
+import {
+  ensureCorrectTyping,
+  sortAndFilterConnectorConfiguration,
+} from './utils/connector_configuration_utils';
 
 type ConnectorConfigurationActions = Pick<
   PostConnectorConfigurationActions,
@@ -75,176 +76,9 @@ export interface CategoryEntry {
 }
 
 export interface ConfigView {
+  advancedConfigurations: ConfigEntryView[];
   categories: CategoryEntry[];
   unCategorizedItems: ConfigEntryView[];
-}
-
-/**
- *
- * Sorts and filters the connector configuration
- *
- * Sorting is done by specified order (if present)
- * otherwise by alphabetic order of keys
- *
- * Filtering is done on any fields with ui_restrictions
- * or that have not had their dependencies met
- *
- */
-function sortAndFilterConnectorConfiguration(
-  config: ConnectorConfiguration,
-  isNative: boolean
-): ConfigView {
-  // This casting is ugly but makes all of the iteration below work for TypeScript
-  // extract_full_html is only defined for crawlers, who don't use this config screen
-  // we explicitly filter it out as well
-  const entries = Object.entries(
-    config as Omit<ConnectorConfiguration, 'extract_full_html'>
-  ).filter(([key]) => key !== 'extract_full_html');
-  const groupedConfigView = entries
-    .map(([key, entry]) => {
-      if (!entry || !isCategoryEntry(entry)) {
-        return null;
-      }
-      const configEntries = entries
-        .map(([configKey, configEntry]) => {
-          if (!configEntry || isCategoryEntry(configEntry) || configEntry.category !== key) {
-            return null;
-          }
-          return { key: configKey, ...configEntry };
-        })
-        .filter(isNotNullish);
-      return { ...entry, configEntries, key };
-    })
-    .filter(isNotNullish);
-
-  const unCategorizedItems = filterSortValidateEntries(
-    entries
-      .map(([key, entry]) =>
-        entry && !isCategoryEntry(entry) && !entry.category ? { key, ...entry } : null
-      )
-      .filter(isNotNullish),
-    config,
-    isNative
-  );
-  const categories = groupedConfigView
-    .map((category) => {
-      const configEntries = filterSortValidateEntries(category.configEntries, config, isNative);
-
-      return configEntries.length > 0 ? { ...category, configEntries } : null;
-    })
-    .filter(isNotNullish);
-
-  return { categories, unCategorizedItems };
-}
-
-function filterSortValidateEntries(
-  configEntries: Array<ConnectorConfigProperties & { key: string }>,
-  config: ConnectorConfiguration,
-  isNative: boolean
-): ConfigEntryView[] {
-  return configEntries
-    .filter(
-      (configEntry) =>
-        ((configEntry.ui_restrictions ?? []).length <= 0 || !isNative) &&
-        dependenciesSatisfied(configEntry.depends_on, config)
-    )
-    .sort((a, b) => {
-      if (isNotNullish(a.order)) {
-        if (isNotNullish(b.order)) {
-          return a.order - b.order;
-        }
-        return -1;
-      }
-      if (isNotNullish(b.order)) {
-        // a doesn't have an order, but b has an order so takes precedence
-        return 1;
-      }
-      return a.key.localeCompare(b.key);
-    })
-    .map((configEntry) => {
-      const label = configEntry.label;
-
-      const validationErrors = [];
-
-      if (configEntry.type === FieldType.INTEGER && !validIntInput(configEntry.value)) {
-        validationErrors.push(
-          i18n.translate(
-            'xpack.enterpriseSearch.content.indices.configurationConnector.config.invalidInteger',
-            {
-              defaultMessage: '{label} must be an integer.',
-              values: { label },
-            }
-          )
-        );
-      }
-
-      return {
-        ...configEntry,
-        is_valid: validationErrors.length <= 0,
-        validation_errors: validationErrors,
-      };
-    });
-}
-
-function validIntInput(value: string | number | boolean | null): boolean {
-  // reject non integers (including x.0 floats), but don't validate if empty
-  return (value !== null || value !== '') &&
-    (isNaN(Number(value)) ||
-      !Number.isSafeInteger(Number(value)) ||
-      ensureStringType(value).indexOf('.') >= 0)
-    ? false
-    : true;
-}
-
-function ensureCorrectTyping(
-  type: FieldType,
-  value: string | number | boolean | null
-): string | number | boolean | null {
-  switch (type) {
-    case FieldType.INTEGER:
-      return validIntInput(value) ? ensureIntType(value) : value;
-    case FieldType.BOOLEAN:
-      return ensureBooleanType(value);
-    default:
-      return ensureStringType(value);
-  }
-}
-
-export function ensureStringType(value: string | number | boolean | null): string {
-  return value !== null ? String(value) : '';
-}
-
-export function ensureIntType(value: string | number | boolean | null): number | null {
-  // int is null-safe to prevent empty values from becoming zeroes
-  if (value === null || value === '') {
-    return null;
-  }
-
-  return parseInt(String(value), 10);
-}
-
-export function ensureBooleanType(value: string | number | boolean | null): boolean {
-  return Boolean(value);
-}
-
-export function dependenciesSatisfied(
-  dependencies: Dependency[],
-  dependencyLookup: ConnectorConfiguration
-): boolean {
-  if (!dependencies) {
-    return true;
-  }
-
-  for (const dependency of dependencies) {
-    // casting here because this is always going to be a ConnectorConfigProperties and not a Category
-    if (
-      dependency.value !== (dependencyLookup[dependency.field] as ConnectorConfigProperties)?.value
-    ) {
-      return false;
-    }
-  }
-
-  return true;
 }
 
 export const ConnectorConfigurationLogic = kea<
